ZDTをpymooで得た正解値を使ってテスト

+++ src/Problems/ZDT1.cpp
@@ -11,7 +11,7 @@ double ZDT1::F1(double x1) const {
 
 double ZDT1::G(const Eigen::ArrayXd& X) const {
     double sum = X.sum();
-    return 1.0 + 9.0 * sum / (DecisionVariablesNum() - 1);
+    return 1.0 + 9.0 * sum / (static_cast<double>(DecisionVariablesNum() - 1));
 }
 
 double ZDT1::H(double f1, double g) const {
